+++ packages/cli/src/types.ts
@@ -10,6 +10,7 @@ declare global {
       IMA_CLI_LAZY_SERVER?: string;
       IMA_CLI_DEV_SERVER_PUBLIC_URL?: string;
       IMA_CLI_PUBLIC_PATH?: string;
+      IMA_CLI_OPEN?: string;
     }
   }
 }