+++ packages/core/src/router/__tests__/ClientRouterSpec.ts
@@ -1,3 +1,7 @@
+/**
+ * @jest-environment jsdom
+ */
+
 import Dispatcher from '../../event/Dispatcher';
 import PageManager from '../../page/manager/PageManager';
 import ActionTypes from '../ActionTypes';