+++ packages/react-page-renderer/src/renderer/__tests__/ClientPageRendererSpec.ts
@@ -181,8 +181,9 @@ describe('ClientPageRenderer', () => {
       expect(pageRenderer['_updateMetaAttributes']).toHaveBeenCalled();
     });
 
-    it('should return resolved promise with object of property content, status and pageState', async () => {
+    it('should return resolved promise with object of property status and pageState', async () => {
       jest.spyOn(controller, 'getHttpStatus').mockReturnValue(200);
+      jest.spyOn(controller, 'getState').mockReturnValue(params);
 
       const response = await pageRenderer.mount(
         controller,
@@ -192,10 +193,7 @@ describe('ClientPageRenderer', () => {
       );
 
       expect(response).toStrictEqual({
-        pageState: {
-          param1: 'param1',
-          param2: 'param2',
-        },
+        pageState: params,
         status: 200,
       });
     });
@@ -257,8 +255,9 @@ describe('ClientPageRenderer', () => {
       expect(pageRenderer['_updateMetaAttributes']).toHaveBeenCalled();
     });
 
-    it('should return resolved promise with object of property content, status and pageState', async () => {
+    it('should return resolved promise with object of property status and pageState', async () => {
       jest.spyOn(controller, 'getHttpStatus').mockReturnValue(200);
+      jest.spyOn(controller, 'getState').mockReturnValue(params);
 
       const response = await pageRenderer.update(
         controller,
@@ -267,6 +266,7 @@ describe('ClientPageRenderer', () => {
       );
 
       expect(response).toStrictEqual({
+        pageState: params,
         status: 200,
       });
     });