+++ packages/server/lib/factory/__tests__/serverAppFactorySpec.js
@@ -1,3 +1,10 @@
+const {
+  GenericError,
+  ServerRouter,
+  Cache,
+  PageStateManager,
+} = require('@ima/core');
+const { toMockedInstance } = require('to-mock');
 const serverAppFactory = require('../serverAppFactory.js');
 const { Emitter, Event } = require('../../emitter.js');
 const { createMonitoring } = require('@esmj/monitor');
@@ -92,8 +99,7 @@ describe('Server App Factory', () => {
     languageLoader = jest.fn();
     applicationFolder = '';
 
-    // TODO IMA@18+ rewrite with toMockInstance
-    router = {
+    router = toMockedInstance(ServerRouter, {
       isClientError: () => false,
       isRedirection: () => false,
       handleError: () =>
@@ -101,30 +107,26 @@ describe('Server App Factory', () => {
           status: 500,
           content: '500 app html',
         }),
-      init: jest.fn(),
-      route: jest.fn(),
-      getPath: jest.fn(),
       getCurrentRouteInfo: jest.fn(() => {
         return {
           route: {
             getName() {
-              return 'notfound';
+              return 'notFound';
             },
           },
         };
       }),
-    };
+    });
 
-    cache = {
+    cache = toMockedInstance(Cache, {
       serialize: jest
         .fn()
         .mockReturnValue(JSON.stringify({ cacheKey: 'cacheValue' })),
-      clear: jest.fn(),
-    };
+    });
 
-    pageStateManager = {
+    pageStateManager = toMockedInstance(PageStateManager, {
       getState: jest.fn().mockReturnValue({ page: 'state' }),
-    };
+    });
 
     appFactory = jest.fn(() => {
       return {
@@ -190,6 +192,7 @@ describe('Server App Factory', () => {
       status: jest.fn(),
       send: jest.fn(),
       set: jest.fn(),
+      redirect: jest.fn(),
       locals: {},
       headerSent: false,
     };
@@ -303,20 +306,6 @@ describe('Server App Factory', () => {
       expect(appFactory).not.toHaveBeenCalled();
     });
 
-    // TODO IMA@18 need performance test for usefulness
-    // it('should render SPA page with cache', async () => {
-    //   jest
-    //     .spyOn(instanceRecycler, 'hasReachedMaxConcurrentRequests')
-    //     .mockReturnValue(true);
-
-    //   await serverApp.requestHandlerMiddleware(REQ, RES);
-    //   const page = await serverApp.requestHandlerMiddleware(REQ, RES);
-
-    //   expect(page.SPA).toBeTruthy();
-    //   expect(page.status).toEqual(200);
-    //   expect(page.cache).toBeTruthy();
-    // });
-
     it('should render overloaded message', async () => {
       environment.$Server.overloadConcurrency = 0;
 
@@ -353,6 +342,58 @@ describe('Server App Factory', () => {
       expect(response.cache).toBeFalsy();
       expect(response.static).toBeFalsy();
       expect(response.content).toBe('404 page');
+    });
+
+    it('should redirect page with 301 status for exceed staticConcurrency', async () => {
+      jest.spyOn(router, 'route').mockReturnValue(
+        Promise.reject(
+          new GenericError('Redirect', {
+            status: 301,
+            url: 'https://imajs.io',
+          })
+        )
+      );
+      jest.spyOn(router, 'isRedirection').mockReturnValue(true);
+      jest.spyOn(router, 'getCurrentRouteInfo').mockReturnValue({
+        route: {
+          getName() {
+            return 'home ';
+          },
+        },
+      });
+
+      environment.$Server.staticConcurrency = 0;
+
+      const response = await serverApp.requestHandlerMiddleware(REQ, RES);
+
+      expect(response.SPA).toBeFalsy();
+      expect(response.status).toBe(301);
+      expect(response.cache).toBeFalsy();
+      expect(response.static).toBeFalsy();
+      expect(response.content).toBeNull();
+      expect(RES.redirect).toHaveBeenCalled();
+    });
+
+    it('should redirect page with 301 status for not exceed staticConcurrency', async () => {
+      jest.spyOn(router, 'route').mockReturnValue(
+        Promise.reject(
+          new GenericError('Redirect', {
+            status: 301,
+            url: 'https://imajs.io',
+          })
+        )
+      );
+      jest.spyOn(router, 'isRedirection').mockReturnValue(true);
+      environment.$Server.staticConcurrency = 100;
+
+      const response = await serverApp.requestHandlerMiddleware(REQ, RES);
+
+      expect(response.SPA).toBeFalsy();
+      expect(response.status).toBe(301);
+      expect(response.cache).toBeFalsy();
+      expect(response.static).toBeFalsy();
+      expect(response.content).toBeNull();
+      expect(RES.redirect).toHaveBeenCalled();
     });
 
     it('should preventDefaulted Event.Request hooks', async () => {