+++ types/ima__server/emitter.d.ts
@@ -1,12 +1,10 @@
 export enum Event {
-  Error = '@esmj/emitter.event.error',
-
   BeforeError = 'ima.server.beforeError',
+  Error = 'ima.server.error',
   AfterError = 'ima.server.afterError',
 
   BeforeRequest = 'ima.server.beforeRequest',
   Request = 'ima.server.request',
-  RequestError = 'ima.server.requestError',
   AfterRequest = 'ima.server.afterRequest',
   BeforeResponse = 'ima.server.beforeResponse',
   Response = 'ima.server.response',